Re-organize project
