+++ 2023/dayone/input.txt
