+++ 2023/main.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"AoC2023/dayone"
+	"AoC2023/utils"
+)
+
+func main() {
+	utils.PrintAnswer(1, 1, dayone.Part1("dayone/input.txt"))
+	utils.PrintAnswer(1, 2, dayone.Part2("dayone/input.txt"))
+}
